Fix PortfolioMapper: restore calculateGrossAmount(BigDecimal, BigDecimal) overload

+++ tradepulseai-backend/cust-service/src/main/java/com/tradepulseai/custservice/mapper/PortfolioMapper.java
@@ -106,8 +106,8 @@ public class PortfolioMapper {
         return response;
     }
 
-    public static BigDecimal calculateGrossAmount(BigDecimal price, int quantity) {
-        return scaleMoney(price.multiply(BigDecimal.valueOf(quantity)));
+    public static BigDecimal calculateGrossAmount(BigDecimal price, BigDecimal quantity) {
+        return scaleMoney(price.multiply(quantity));
     }
